Revert "move tests to folder web3mail"
This reverts commit 5f3792c715b920ce5d9b2c65e40c1956e7aae88d.
